run_node: run a D3.t on any Dom.node
The previous interface for the run operation was too restrictive, as it
required D3 operations to run on trees that were attached to the
document. With this interface, it is now possible to run D3 operations
on trees that are not attached, for example those that have been
generated using tyxml.

Related to #1, #2.