datum: fix type of datum operator
The datum operator does not introduce a data bind as the data operator
does. It simply transforms the existing data that's attached to a
node, and should therefore have a non-list return type. If the data
attached to a node happens to be a list, that's ok given the new type.